Refactor to use a command and handler
