+++ config/mail.php
@@ -28,18 +28,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | The destination email address.
-    |--------------------------------------------------------------------------
-    |
-    | Here you should provide the email address all your notifications should
-    | be sent to.
-    |
-    */
-
-    'destination' => 'user@example.com',
-
-    /*
-    |--------------------------------------------------------------------------
     | SMTP Host Address
     |--------------------------------------------------------------------------
     |
